Cue phrase and stigma phrase feature added and term weight fixed
features - Added phrase_feature that can evaluate a list of words in every sentence. (Cue phrase and Stigma words use this feature)
                - Fixed keyword feature but values still over 1
text-summarizer - Added resource_loader to load every file in resource and save it as a list of words, also added "constants" for the bonus_words and stigma_words names
                              - Added the feature value in the main
textClasses - Class word has now term weight as a property and the setter is also defined.